refactor: split the provider health tracker from its vocabulary
The record, summary and status a health surface speaks are one concern;
accumulating outcomes under a lock with a memory bound is another. Give
the second its own module so neither has to be read to understand the
first.
